Fix compatibility issue with Apache2 websocket proxy
The Apache2 proxy uses "Upgrade: WebSocket" headers when forwarding a
WebSocket connection (note the camel case), breaking interoperability
with the web server.

The WebSocket RFC 6455 is somewhat inconclusive as it uses "Upgrade:
websocket" in all its examples and requires the "websocket" keyword in
section 4.1, but states that "WebSocket" has been registered as HTTP
Upgrade Token in section 11.2.

In any case, following the robustness principle to be as liberal as
possible when accepting data and as conservative as possible when
sending it, this commit makes the "websocket" keyword matching case
insensitive in the request handler.